+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -86,6 +86,7 @@ public let masterRuleList = RuleList(rules: [
     LegacyHashingRule.self,
     LegacyMultipleRule.self,
     LegacyNSGeometryFunctionsRule.self,
+    LegacyObjcTypeRule.self,
     LegacyRandomRule.self,
     LetVarWhitespaceRule.self,
     LineLengthRule.self,