+++ Tests/LinuxMain.swift
@@ -790,6 +790,12 @@ extension LegacyNSGeometryFunctionsRuleTests {
     ]
 }
 
+extension LegacyObjcTypeRuleTests {
+    static var allTests: [(String, (LegacyObjcTypeRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
 extension LegacyRandomRuleTests {
     static var allTests: [(String, (LegacyRandomRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
@@ -1755,6 +1761,7 @@ XCTMain([
     testCase(LegacyHashingRuleTests.allTests),
     testCase(LegacyMultipleRuleTests.allTests),
     testCase(LegacyNSGeometryFunctionsRuleTests.allTests),
+    testCase(LegacyObjcTypeRuleTests.allTests),
     testCase(LegacyRandomRuleTests.allTests),
     testCase(LetVarWhitespaceRuleTests.allTests),
     testCase(LineEndingTests.allTests),